Move `torch.linalg` opinfos to opinfo.definitions
[ghstack-poisoned]
